Refactor service_worker_jobs.cc and move job related function into service_worker_abstract_jobs.cc
b/286597506
