+++ cobalt/worker/service_worker_jobs.h
@@ -42,6 +42,7 @@
 #include "cobalt/worker/client_query_options.h"
 #include "cobalt/worker/frame_type.h"
 #include "cobalt/worker/service_worker.h"
+#include "cobalt/worker/service_worker_abstract_jobs.h"
 #include "cobalt/worker/service_worker_consts.h"
 #include "cobalt/worker/service_worker_object.h"
 #include "cobalt/worker/service_worker_registration.h"
@@ -58,144 +59,14 @@ namespace worker {
 
 // Algorithms for Service Worker Jobs.
 //   https://www.w3.org/TR/2022/CRD-service-workers-20220712/#algorithms
-class ServiceWorkerJobs {
+class ServiceWorkerJobs : public ServiceWorkerAbstractJobs {
  public:
-  // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#dfn-job-type
-  enum JobType { kRegister, kUpdate, kUnregister };
-
-  class JobQueue;
-
-  // This type handles the different promise variants used in jobs.
-  class JobPromiseType {
-   public:
-    // Constructors for each promise variant that can be held.
-    explicit JobPromiseType(
-        std::unique_ptr<script::ValuePromiseBool::Reference> promise_reference);
-    explicit JobPromiseType(
-        std::unique_ptr<script::ValuePromiseWrappable::Reference>
-            promise_reference);
-
-    template <typename PromiseReference>
-    static std::unique_ptr<JobPromiseType> Create(
-        PromiseReference promise_reference) {
-      return std::unique_ptr<JobPromiseType>(
-          new JobPromiseType(std::move(promise_reference)));
-    }
-
-    void Resolve(const bool result);
-    void Resolve(const scoped_refptr<cobalt::script::Wrappable>& result);
-    void Reject(script::SimpleExceptionType exception);
-    void Reject(web::DOMException::ExceptionCode code,
-                const std::string& message);
-    void Reject(const scoped_refptr<script::ScriptException>& result);
-
-    bool is_pending() const { return is_pending_.load(); }
-
-   private:
-    starboard::atomic_bool is_pending_{true};
-    std::unique_ptr<script::ValuePromiseBool::Reference>
-        promise_bool_reference_;
-    std::unique_ptr<script::ValuePromiseWrappable::Reference>
-        promise_wrappable_reference_;
-  };
-
-  // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#dfn-job
-  struct Job {
-    Job(JobType type, const url::Origin& storage_key, const GURL& scope_url,
-        const GURL& script_url, web::Context* client,
-        std::unique_ptr<JobPromiseType> promise)
-        : type(type),
-          storage_key(storage_key),
-          scope_url(scope_url),
-          script_url(script_url),
-          update_via_cache(
-              ServiceWorkerUpdateViaCache::kServiceWorkerUpdateViaCacheImports),
-          client(client),
-          promise(std::move(promise)) {}
-    ~Job() {
-      client = nullptr;
-      containing_job_queue = nullptr;
-    }
-
-    // Job properties from the spec.
-    //
-    JobType type;
-    url::Origin storage_key;
-    GURL scope_url;
-    GURL script_url;
-    ServiceWorkerUpdateViaCache update_via_cache;
-    web::Context* client;
-    GURL referrer;
-    std::unique_ptr<JobPromiseType> promise;
-    JobQueue* containing_job_queue = nullptr;
-    std::deque<std::unique_ptr<Job>> equivalent_jobs;
-    bool force_bypass_cache_flag = false;
-    bool no_promise_okay = false;
-
-    // Custom, not in the spec.
-    //
-
-    // This lock is for the list of equivalent jobs. It should also be held when
-    // resolving the promise.
-    base::Lock equivalent_jobs_promise_mutex;
-
-    // The loader that is used for asynchronous loads.
-    std::unique_ptr<loader::Loader> loader;
-  };
-
-  // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#dfn-job-queue
-  class JobQueue {
-   public:
-    bool empty() {
-      base::AutoLock lock(mutex_);
-      return jobs_.empty();
-    }
-    void Enqueue(std::unique_ptr<Job> job) {
-      base::AutoLock lock(mutex_);
-      jobs_.push_back(std::move(job));
-    }
-    std::unique_ptr<Job> Dequeue() {
-      base::AutoLock lock(mutex_);
-      std::unique_ptr<Job> job;
-      job.swap(jobs_.front());
-      jobs_.pop_front();
-      return job;
-    }
-    Job* FirstItem() {
-      base::AutoLock lock(mutex_);
-      return jobs_.empty() ? nullptr : jobs_.front().get();
-    }
-
-    // Also return a held autolock, to ensure the item remains a valid item in
-    // the queue while it's in use.
-    std::pair<Job*, base::sequence_manager::MoveableAutoLock> LastItem() {
-      base::sequence_manager::MoveableAutoLock lock(mutex_);
-      Job* job = jobs_.empty() ? nullptr : jobs_.back().get();
-      return std::pair<Job*, base::sequence_manager::MoveableAutoLock>(
-          job, std::move(lock));
-    }
-
-    // Ensure no references are kept to JS objects for a client that is about to
-    // be shutdown.
-    void PrepareForClientShutdown(web::Context* client);
-
-    // Helper method for PrepareForClientShutdown to help with recursion to
-    // equivalent jobs.
-    void PrepareJobForClientShutdown(const std::unique_ptr<Job>& job,
-                                     web::Context* client);
-
-   private:
-    base::Lock mutex_;
-    std::deque<std::unique_ptr<Job>> jobs_;
-  };
-
   ServiceWorkerJobs(web::WebSettings* web_settings,
                     network::NetworkModule* network_module,
                     web::UserAgentPlatformInfo* platform_info,
                     base::MessageLoop* message_loop, const GURL& url);
-  ~ServiceWorkerJobs();
+  ~ServiceWorkerJobs() override;
 
-  base::MessageLoop* message_loop() { return message_loop_; }
 
   // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#start-register-algorithm
   void StartRegister(const base::Optional<GURL>& scope_url,
@@ -271,11 +142,7 @@ class ServiceWorkerJobs {
   // Registration of web contexts that may have service workers.
   void RegisterWebContext(web::Context* context);
   void UnregisterWebContext(web::Context* context);
-  bool IsWebContextRegistered(web::Context* context) {
-    DCHECK(base::MessageLoop::current() == message_loop());
-    return web_context_registrations_.end() !=
-           web_context_registrations_.find(context);
-  }
+
 
   // Ensure no references are kept to JS objects for a client that is about to
   // be shutdown.
@@ -284,23 +151,7 @@ class ServiceWorkerJobs {
   // Set the active worker for a client if there is a matching service worker.
   void SetActiveWorker(web::EnvironmentSettings* client);
 
-  // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#create-job
-  std::unique_ptr<Job> CreateJob(
-      JobType type, const url::Origin& storage_key, const GURL& scope_url,
-      const GURL& script_url,
-      std::unique_ptr<script::ValuePromiseWrappable::Reference> promise,
-      web::Context* client) {
-    return CreateJob(type, storage_key, scope_url, script_url,
-                     JobPromiseType::Create(std::move(promise)), client);
-  }
-  std::unique_ptr<Job> CreateJob(
-      JobType type, const url::Origin& storage_key, const GURL& scope_url,
-      const GURL& script_url,
-      std::unique_ptr<script::ValuePromiseBool::Reference> promise,
-      web::Context* client) {
-    return CreateJob(type, storage_key, scope_url, script_url,
-                     JobPromiseType::Create(std::move(promise)), client);
-  }
+
   std::unique_ptr<Job> CreateJobWithoutPromise(JobType type,
                                                const url::Origin& storage_key,
                                                const GURL& scope_url,
@@ -310,13 +161,6 @@ class ServiceWorkerJobs {
     job->no_promise_okay = true;
     return job;
   }
-  std::unique_ptr<Job> CreateJob(
-      JobType type, const url::Origin& storage_key, const GURL& scope_url,
-      const GURL& script_url, std::unique_ptr<JobPromiseType> promise = nullptr,
-      web::Context* client = nullptr);
-
-  // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#schedule-job
-  void ScheduleJob(std::unique_ptr<Job> job);
 
   // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#activation-algorithm
   void Activate(ServiceWorkerRegistrationObject* registration);
@@ -355,41 +199,10 @@ class ServiceWorkerJobs {
     bool has_updated_resources = false;
   };
 
-  // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#dfn-scope-to-job-queue-map
-  using JobQueueMap = std::map<std::string, std::unique_ptr<JobQueue>>;
-
-  // Type to hold the errorData for rejection of promises.
-  class PromiseErrorData {
-   public:
-    explicit PromiseErrorData(const script::MessageType& message_type)
-        : message_type_(message_type),
-          exception_code_(web::DOMException::kNone) {}
-    PromiseErrorData(const web::DOMException::ExceptionCode& code,
-                     const std::string& message)
-        : message_type_(script::kNoError),
-          exception_code_(code),
-          message_(message) {}
-
-    void Reject(std::unique_ptr<JobPromiseType> promise) const;
-
-   private:
-    // Use script::MessageType because it can hold kNoError value to distinguish
-    // between simple exceptions and DOM exceptions.
-    script::MessageType message_type_;
-    const web::DOMException::ExceptionCode exception_code_;
-    const std::string message_;
-  };
-
   enum RegistrationState { kInstalling, kWaiting, kActive };
 
-  // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#dfn-job-equivalent
-  bool ReturnJobsAreEquivalent(Job* one, Job* two);
-
-  // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#run-job-algorithm
-  void RunJob(JobQueue* job_queue);
-
   // Task for "Run Job" to run in the service worker thread.
-  void RunJobTask(JobQueue* job_queue);
+  void RunJobTask(JobQueue* job_queue) override;
 
   // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#register-algorithm
   void Register(Job* job);
@@ -413,20 +226,6 @@ class ServiceWorkerJobs {
   // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#unregister-algorithm
   void Unregister(Job* job);
 
-  // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#reject-job-promise
-  void RejectJobPromise(Job* job, const PromiseErrorData& error_data);
-
-  // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#resolve-job-promise-algorithm
-  void ResolveJobPromise(
-      Job* job, const scoped_refptr<ServiceWorkerRegistrationObject>& value) {
-    ResolveJobPromise(job, false, value);
-  }
-  void ResolveJobPromise(Job* job, bool value,
-                         const scoped_refptr<ServiceWorkerRegistrationObject>&
-                             registration = nullptr);
-
-  // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#finish-job-algorithm
-  void FinishJob(Job* job);
 
   // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#run-service-worker-algorithm
   // The return value is a 'Completion or failure'.
@@ -443,9 +242,6 @@ class ServiceWorkerJobs {
 
   // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#try-activate-algorithm
   void TryActivate(ServiceWorkerRegistrationObject* registration);
-
-  // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#service-worker-has-no-pending-events
-  bool ServiceWorkerHasNoPendingEvents(ServiceWorkerObject* worker);
 
   // https://www.w3.org/TR/2022/CRD-service-workers-20220712/#update-registration-state-algorithm
   void UpdateRegistrationState(
@@ -478,16 +274,7 @@ class ServiceWorkerJobs {
   std::unique_ptr<loader::FetcherFactory> fetcher_factory_;
   // LoaderFactory that is used to acquire references to resources from a URL.
   std::unique_ptr<loader::ScriptLoaderFactory> script_loader_factory_;
-  base::MessageLoop* message_loop_;
-
-  JobQueueMap job_queue_map_;
   std::unique_ptr<ServiceWorkerRegistrationMap> scope_to_registration_map_;
-
-  std::set<web::Context*> web_context_registrations_;
-
-  base::WaitableEvent web_context_registrations_cleared_ = {
-      base::WaitableEvent::ResetPolicy::MANUAL,
-      base::WaitableEvent::InitialState::NOT_SIGNALED};
 };
 
 }  // namespace worker
